package-managers: moved npm package manager to separete file
